Scanner-Api/requirements.txt
flask
tensorflow
flask-cors
pillow
gunicorn

